service/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels-auth-token-middlewares==1.1.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
elastic-transport==8.15.1
elasticsearch==8.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.2
tika==2.6.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.2
pillow==11.2.1
